requirements.txt
flask
markdown
python-dotenv